Mushroom poisonous binary classification
In this project, we aim to build a machine learning model to classify mushrooms as either poisonous or edible. This classification is crucial for safety and public health as consuming poisonous mushrooms can be dangerous. We will use a dataset from Kaggle that contains various features of mushroom species to build our model.